+++ CHANGELOG.md
@@ -5,10 +5,14 @@
   * New callback: `onSort`.
   * New attribute: `ignoreSort`.
     * Ignore column sorting functions to maintain the natural order of the data. Allows the header to be sortable if the column is also sortable.
+* `Davi`
+  * New attribute: `tapToSortEnabled`.
+    * Indicates whether sorting events are enabled on the header.
+final bool sortable;
 * Refactor
   * The `sortable` attribute of the `DaviColumn` can be `TRUE` even without a `sort` function.
   * Typedef `DaviColumnSort`
-    * Attribute addition: `DaviColumn<DATA> column`
+    * New parameter: `DaviColumn<DATA> column`
 
 ## 3.0.0